refactor: rename folder, state inline banner deps
