+++ docs/public/__registry__/component/index.json
@@ -55,9 +55,13 @@
   },
   {
     "name": "inline-banner",
-    "dependencies": [],
+    "dependencies": [
+      "@radix-ui/react-slot",
+      "@daangn/react-monochrome-icon"
+    ],
     "files": [
-      "component/inline-banner.tsx"
+      "component/inline-banner.tsx",
+      "hook/use-dismissible.ts"
     ]
   },
   {
