+++ docs/registry/hook/use-dismissible.ts
